+++ indico/modules/events/client/js/creation.js
@@ -207,7 +207,6 @@ import {camelizeKeys} from 'indico/utils/case';
         !('room_id' in roomData) ||
         !startDt.isValid() ||
         !endDt.isValid() ||
-        !startDt.isSame(endDt, 'day') ||
         startDt.isSameOrAfter(endDt) ||
         isCategoryExcluded(category['id']) ||
         timezone !== options.serverDefaultTimezone ||