+++ indico/modules/events/sessions/models/blocks.py
@@ -19,6 +19,7 @@ from indico.modules.events.timetable.models.entries import TimetableEntry
 from indico.util.iterables import materialize_iterable
 from indico.util.locators import locator_property
 from indico.util.string import format_repr, slugify
+from indico.web.flask.util import url_for
 
 
 class SessionBlock(LocationMixin, db.Model):
@@ -72,7 +73,7 @@ class SessionBlock(LocationMixin, db.Model):
     # relationship backrefs:
     # - contributions (Contribution.session_block)
     # - legacy_mapping (LegacySessionBlockMapping.session_block)
-    # - room_reservation_links (ReservationLink.session_block)
+    # - room_reservation_occurrence_links (ReservationOccurrenceLink.session_block)
     # - session (Session.blocks)
     # - timetable_entry (TimetableEntry.session_block)
     # - vc_room_associations (VCRoomEventAssociation.linked_block)
@@ -119,6 +120,10 @@ class SessionBlock(LocationMixin, db.Model):
     @property
     def full_title(self):
         return f'{self.session.title}: {self.title}' if self.title else self.session.title
+
+    @property
+    def url(self):
+        return url_for('sessions.display_session', self)
 
     def can_manage(self, user, allow_admin=True):
         return self.session.can_manage_blocks(user, allow_admin=allow_admin)